requirements.txt
numpy>=2.2.4
pyzmq>=26.3.0
# pygobject>=3.52.3
sqlalchemy>=2.0.35
botocore>=1.35.0
